feat: add embedding utilities
