+++ split_clusters.py
@@ -9,72 +9,19 @@ import argparse
 import csv
 import json
 from pathlib import Path
-from typing import List, Tuple, Dict
+from typing import Dict, List
 
 import numpy as np
-from PIL import Image
-
-import torch
-from facenet_pytorch import InceptionResnetV1
 import joblib
 
 from config import get_profile
+from embedding_utils import embed_images, get_device, load_images
 
 IMAGE_EXTS = {'.jpg', '.jpeg', '.png', '.bmp', '.tif', '.tiff', '.webp'}
 
+
 def list_images(root: Path) -> List[Path]:
     return sorted([p for p in root.rglob('*') if p.is_file() and p.suffix.lower() in IMAGE_EXTS])
-
-def batched(iterable, n):
-    batch = []
-    for x in iterable:
-        batch.append(x)
-        if len(batch) == n:
-            yield batch
-            batch = []
-    if batch:
-        yield batch
-
-def load_images(paths: List[Path]) -> List[Image.Image]:
-    out = []
-    for p in paths:
-        try:
-            out.append(Image.open(p).convert("RGB"))
-        except Exception:
-            out.append(None)
-    return out
-
-def embed_images(imgs: List[Image.Image], device: str, batch_size: int) -> np.ndarray:
-    resnet = InceptionResnetV1(pretrained='vggface2').eval().to(device)
-    embs = np.zeros((len(imgs), 512), dtype=np.float32)
-
-    def preprocess(pil):
-        return pil.resize((160,160))
-
-    idx = 0
-    for chunk in batched(imgs, batch_size):
-        good_idx = []
-        tensors = []
-        for j, im in enumerate(chunk):
-            if im is None:
-                continue
-            try:
-                im2 = preprocess(im)
-                t = torch.from_numpy(np.asarray(im2)).permute(2,0,1).float() / 255.0
-                tensors.append(t.unsqueeze(0))
-                good_idx.append(j)
-            except Exception:
-                pass
-        if not tensors:
-            idx += len(chunk)
-            continue
-        batch = torch.cat(tensors, dim=0).to(device)
-        with torch.no_grad():
-            feats = resnet(batch).cpu().numpy().astype(np.float32)
-        for j_local, vec in zip(good_idx, feats):
-            embs[idx + j_local, :] = vec
-        idx += len(chunk)
-    return embs
 
 def main():
     parser = argparse.ArgumentParser(description="Predict faces for a folder of images/crops")
@@ -88,15 +35,7 @@ def main():
 
     prof = get_profile(args.strictness)
 
-    # Device selection
-    device = args.device
-    if device is None:
-        if torch.cuda.is_available():
-            device = "cuda"
-        elif getattr(torch.backends, "mps", None) and torch.backends.mps.is_available():
-            device = "mps"
-        else:
-            device = "cpu"
+    device = get_device(args.device)
     print(f"[INFO] Using device: {device} | embed_batch={prof.embed_batch}")
 
     model_dir = Path(args.model_dir).expanduser().resolve()
